Finally have a fully working version of pause time. It has inbuilt limitations required by this system:  * No shooting people into pause time. More generally no interacting in any way with paused guys whatsoever.
As a further reminder: FUTURE SELVES, DON"T IMPLEMENT MORE PAUSE TIME MANIPULATION.

Shooting people into pause time would most likely break the assertion that there is always 1 player frame as remote departure edits take time to propagate through yet do not change relative index.

Anyway have a fit with this version and test bugs, many things are untested because I can't be bothered writing replays to test it. Hold 'P' to test the engine is still in this version. Removing it soon to work on other things.

+++ Guy.cpp
@@ -14,7 +14,7 @@ Guy::Guy(int nX,
          int nBoxPauseLevel,
          TimeDirection nTimeDirection,
          int nPauseLevel,
-         size_t nRelativeIndex,
+         int nRelativeIndex,
          int nSubimage) :
 referenceCount(new int(1)),
 data(new Data(nX,nY,nXspeed,nYspeed,nWidth,nHeight, nSupported,
